Moved param manager to library in rosplane package
