+++ rosplane_extra/include/input_mapper.hpp
@@ -7,7 +7,7 @@
 #include <rosplane_msgs/msg/controller_commands.hpp>
 #include <rosplane_msgs/msg/state.hpp>
 #include <std_srvs/srv/trigger.hpp>
-#include <param_manager.hpp>
+#include <param_manager/param_manager.hpp>
 
 namespace rosplane
 {
